+++ course/format/topcoll/tests/format_topcoll_courseformatlib_test.php
@@ -24,13 +24,12 @@
  * @link       http://docs.moodle.org/en/Collapsed_Topics_course_format
  * @license    http://www.gnu.org/copyleft/gpl.html GNU Public License
  */
-defined('MOODLE_INTERNAL') || die();
 
 /**
  * Library unit tests for the Collapsed Topics course format.
  * @group format_topcoll
  */
-class format_topcoll_courseformatlib_testcase extends advanced_testcase {
+class format_topcoll_courseformatlib_test extends advanced_testcase {
 
     protected $course;
     protected $courseformat;